Add types for self hosted, start attempting to open purchase modal form the pricing modal.
